requirements.txt
pandas==2.3.3
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
httplib2==0.31.0
python-dotenv==1.2.1

requests 
pypdf==6.5.0


openpyxl==3.1.5